+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "farmbot",
-  "version": "0.1.2",
+  "version": "0.2.0",
   "description": "Javascript wrapper for Farmbot's authentication and RPC instructions",
   "main": "dist/farmbot.js",
   "scripts": {
-    "test": "gulp build; karma start karma.conf.js"
+    "test": "karma start karma.conf.js"
   },
   "repository": {
     "type": "git",
